Fixed missing return type. Updated header imports.
git-svn-id: http://wxmax.googlecode.com/svn/trunk/wx.mod@853 971f8453-5c39-0410-a452-3998bf019487

+++ wxprocess.mod/glue.cpp
@@ -199,7 +199,7 @@ int bmx_wxkill(long pid, wxSignal signal, wxKillError * rc, int flags) {
 	return wxKill(pid, signal, rc, flags);
 }
 
-long bmx_wxexecute(BBString * command, int sync, wxProcess * callback) {
+int bmx_wxexecute(BBString * command, int sync, wxProcess * callback) {
 	if (callback) {
 		return wxExecute(wxStringFromBBString(command), sync, callback);
 	} else {
